+++ .eslintrc.js
@@ -1,3 +1,8 @@
+/**
+ * @file .eslintrc.js
+ * @author James Wang <github.com/cn-wx>
+ */
+
 module.exports = {
   root: true,
 
@@ -8,7 +13,7 @@ module.exports = {
   extends: ['plugin:vue/essential', '@vue/standard'],
 
   rules: {
-    'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
+    'no-console': 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off'
   },