+++ babel.config.js
@@ -1,3 +1,8 @@
+/**
+ * @file babel.config.js
+ * @author James Wang <github.com/cn-wx>
+ */
+
 module.exports = {
   presets: ['@vue/app']
 }