+++ jest.config.js
@@ -1,3 +1,8 @@
+/**
+ * @file jest.config.js
+ * @author James Wang <github.com/cn-wx>
+ */
+
 module.exports = {
   moduleFileExtensions: [
     'js',