+++ src/demo/Demo.vue
@@ -1,53 +1,31 @@
 <template>
   <main>
-    <img alt="Vue logo" src="./assets/logo.png" />
+    <img style="max-width: 80%" alt="Vue logo" src="./assets/logo.png" />
     <h1><a href="https://github.com/cn-wx/vue-hotspot">Vue-Hotspot</a></h1>
     <p class="desc">Hotspot component for Vue.js.</p>
-    <section>
+
+    <section v-for="(item, i) in hotspots" :key="i">
       <div class="container">
-        <div class="text">Editable: {{ hotspotConfig1.editable }} | interactivity: {{ hotspotConfig1.interactivity }}</div>
-        <!-- Hotspot -->
+        <!-- Title -->
+        <div class="text mb-sm">
+          <span class="bold">Example {{ i + 1 }}</span>
+        </div>
+        <!-- Main -->
         <v-hotspot
-          :init-options="hotspotConfig1"
+          :init-options="item"
           @save-data="saveData"
           @after-delete="afterDelete" />
+        <div class="text">
+          Editable: {{ item.editable }} | interactivity: {{ item.interactivity }}
+        </div>
         <!-- Button -->
         <toggle
           :label="'Change Editable'"
-          :hotspot-config="hotspotConfig1"
+          :hotspot-config="item"
           @change-editable="changeEditable" />
       </div>
     </section>
-    <section>
-      <div class="container">
-        <div class="text">Editable: {{ hotspotConfig2.editable }} | interactivity: {{ hotspotConfig2.interactivity }}</div>
-        <!-- Hotspot -->
-        <v-hotspot
-          :init-options="hotspotConfig2"
-          @save-data="saveData"
-          @after-delete="afterDelete" />
-        <!-- Button -->
-        <toggle
-          :label="'Change Editable'"
-          :hotspot-config="hotspotConfig2"
-          @change-editable="changeEditable" />
-      </div>
-    </section>
-    <section>
-      <div class="container">
-        <div class="text">Editable: {{ hotspotConfig3.editable }} | interactivity: {{ hotspotConfig3.interactivity }}</div>
-        <!-- Hotspot -->
-        <v-hotspot
-          :init-options="hotspotConfig3"
-          @save-data="saveData"
-          @after-delete="afterDelete" />
-        <!-- Button -->
-        <toggle
-          :label="'Change Editable'"
-          :hotspot-config="hotspotConfig3"
-          @change-editable="changeEditable" />
-      </div>
-    </section>
+
     <footer>
       <!-- eslint-disable-next-line -->
       <a href="//github.com/cn-wx">@James Wang</a>|<a href="//github.com/cn-wx/vue-hotspot/blob/master/LICENSE">MIT License</a>|<a href="//github.com/cn-wx/vue-hotspot">View on GitHub</a>
@@ -66,29 +44,40 @@ export default {
   },
   data () {
     return {
-      hotspotConfig1: {
-        image: 'https://about.unimelb.edu.au/__data/assets/image/0022/92911/varieties/large.jpg',
-        editable: true,
-        interactivity: 'click'
-      },
-      hotspotConfig2: {
-        image: 'https://about.unimelb.edu.au/__data/assets/image/0013/10840/varieties/large.jpg',
-        editable: false,
-        interactivity: 'hover'
-      },
-      hotspotConfig3: {
-        image: 'https://about.unimelb.edu.au/__data/assets/image/0021/15582/varieties/banner.jpg',
-        editable: false,
-        interactivity: 'hover',
-        data: [
-          { Message: 'A prepopulated hotspot', Title: 'Vue Hotspot 1', x: 33.3973, y: 58.3333 },
-          { Message: 'Another prepopulated hotspot', Title: 'Vue Hotspot 2', x: 53.3973, y: 78.3333 }
-        ],
-        hotspotColor: '#85ce61',
-        messageBoxColor: '#409eff',
-        textColor: '#333',
-        opacity: 0.9
-      }
+      hotspots: [
+        {
+          image: 'https://about.unimelb.edu.au/__data/assets/image/0022/92911/varieties/large.jpg',
+          editable: true,
+          interactivity: 'hover'
+        },
+        {
+          image: 'https://about.unimelb.edu.au/__data/assets/image/0013/10840/varieties/large.jpg',
+          editable: false,
+          interactivity: 'click'
+        },
+        {
+          image: 'https://about.unimelb.edu.au/__data/assets/image/0021/15582/varieties/banner.jpg',
+          editable: false,
+          interactivity: 'hover',
+          data: [
+            { Message: 'A prepopulated hotspot with custom color', Title: 'Vue Hotspot 1', x: 33.3973, y: 58.3333 },
+            { Message: 'Another prepopulated hotspot with custom color', Title: 'Vue Hotspot 2', x: 53.3973, y: 78.3333 },
+            { Message: 'Another prepopulated hotspot with custom color', Title: 'Vue Hotspot 3', x: 73.3973, y: 28.3333 }
+          ],
+          hotspotColor: '#85ce61',
+          messageBoxColor: '#409eff',
+          textColor: '#333'
+        },
+        {
+          image: 'https://about.unimelb.edu.au/__data/assets/image/0025/11599/varieties/medium.jpg',
+          editable: false,
+          interactivity: 'none',
+          data: [
+            { Message: 'A prepopulated hotspot with interactivity: none', Title: 'Vue Hotspot 1', x: 33.3973, y: 68.3333 },
+            { Message: 'Another prepopulated hotspot with interactivity: none', Title: 'Vue Hotspot 2', x: 73.3973, y: 38.3333 }
+          ]
+        }
+      ]
     }
   },
   methods: {
@@ -109,6 +98,18 @@ export default {
 </script>
 
 <style lang="less">
+@small: 1em;
+@medium: 2em;
+@gray: #666666;
+
+.no-margin{margin: 0}
+.mt-sm {margin-top: @small}.mb-sm {margin-bottom: @small}.ml-sm {margin-left: @small}.mr-sm {margin-right: @small}
+.mt-m{margin-top: @medium}.mb-m{margin-bottom: @medium}.ml-m{margin-left: @medium}.mr-m{margin-right: @medium}
+
+.no-padding{padding: 0}
+.pt-sm {padding-top: @small}.pb-sm {padding-bottom: @small}.pl-sm {padding-left: @small}.pr-sm {padding-right: @small}
+.pt-m {padding-top: @medium}.pb-m {padding-bottom: @medium}.pl-m {padding-left: @medium}.pr-m {padding-right: @medium}
+
 html {
   scroll-behavior: smooth;
 }
@@ -116,7 +117,7 @@ body {
   margin: 0;
   padding: 3em 0 0;
   font-family: "Source Sans Pro", "Helvetica Neue", Arial, sans-serif;
-  color: #666;
+  color: @gray;
   text-align: center;
 }
 a {
@@ -156,8 +157,16 @@ button {
   // min-width: calc(40vw + 4em);
   width: 50vw;
   .text {
-    margin-bottom: 1em;
+    &.mb-sm {
+      margin-bottom: 1em;
+    }
     font-size: 1.2em;
+    .bold {
+      font-weight: bold;
+    }
+    .left {
+      float: left;
+    }
   }
 }
 footer {
@@ -182,7 +191,7 @@ footer {
   @media (max-width: 980px) {
     .container {
       width: 100vw;
-      margin: 1em auto;
+      margin: @small auto;
       padding: 0;
       border: none;
       border-radius: 0;