+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+2.0.0
+* Add Vue Composition-API (https://github.com/vuejs/composition-api)
+
+1.1.8
+* Refactor code with composition-api (#68)
+* Performance enhancement for window resize listener
+
 1.1.7
 * Update usage in README.